Fix extracted dialog components
