+++ src/editor/mapConfig.ts
@@ -1,0 +1,10 @@
+interface MapConfig {
+  name: string;
+  width: number;
+  height: number;
+  tileSize: number;
+  location: string;
+  isStartingMap?: boolean;
+}
+
+export type { MapConfig };
